使用Pydantic重构逻辑
