+++ src/nonebot_plugin_suggarchat/handle/chat.py
@@ -6,6 +6,7 @@ import sys
 import time
 import traceback
 from datetime import datetime
+from typing import Any
 
 from nonebot import logger
 from nonebot.adapters.onebot.v11 import (
@@ -26,19 +27,20 @@ from ..config import config_manager
 from ..event import ChatEvent
 from ..exception import CancelException
 from ..matcher import MatcherManager
-from ..utils import (
-    get_chat,
-    get_current_datetime_timestamp,
-    get_friend_info,
-    get_memory_data,
-    hybrid_token_count,
-    remove_think_tag,
+from ..utils.admin import (
     send_to_admin,
     send_to_admin_as_error,
+)
+from ..utils.functions import (
+    get_current_datetime_timestamp,
+    get_friend_name,
+    remove_think_tag,
     split_message_into_chats,
     synthesize_message,
-    write_memory_data,
 )
+from ..utils.libchat import get_chat
+from ..utils.memory import MemoryModel, get_memory_data, write_memory_data
+from ..utils.tokenizer import hybrid_token_count
 
 
 async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
@@ -50,7 +52,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
         event: GroupMessageEvent,
         matcher: Matcher,
         bot: Bot,
-        group_data: dict,
+        group_data: MemoryModel,
         memory_length_limit: int,
         Date: str,
     ):
@@ -65,7 +67,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
         if not config_manager.config.enable_group_chat:
             matcher.skip()
 
-        if not group_data["enable"]:
+        if not group_data.enable:
             await matcher.send("聊天没有启用")
             return
 
@@ -117,7 +119,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
         else:
             text = event.message.extract_plain_text()
 
-        group_data["memory"]["messages"].append(
+        group_data.memory.messages.append(
             {
                 "role": "user",
                 "content": text,
@@ -138,7 +140,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
         response = await process_chat(event, send_messages, tokens)
 
         # 记录模型回复
-        group_data["memory"]["messages"].append(
+        group_data.memory.messages.append(
             {
                 "role": "assistant",
                 "content": str(response),
@@ -153,7 +155,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
         event: PrivateMessageEvent,
         matcher: Matcher,
         bot: Bot,
-        private_data: dict,
+        private_data: MemoryModel,
         memory_length_limit: int,
         Date: str,
     ):
@@ -192,7 +194,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
                 [
                     {
                         "type": "text",
-                        "text": f"{Date}{await get_friend_info(event.user_id, bot=bot)}（{event.user_id}）： {content!s}",
+                        "text": f"{Date}{await get_friend_name(event.user_id, bot=bot)}（{event.user_id}）： {content!s}",
                     },
                 ]
                 + [
@@ -201,11 +203,11 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
                     if seg.data.get("type") == "image"
                 ]
                 if is_multimodal
-                else f"{Date}{await get_friend_info(event.user_id, bot=bot)}（{event.user_id}）： {content!s}"
+                else f"{Date}{await get_friend_name(event.user_id, bot=bot)}（{event.user_id}）： {content!s}"
             )
         else:
             text = event.message.extract_plain_text()
-        private_data["memory"]["messages"].append({"role": "user", "content": text})
+        private_data.memory.messages.append({"role": "user", "content": text})
         if chat_manager.debug:
             logger.debug(f"当前私聊提示词：\n{config_manager.private_train}")
         # 控制记忆长度和 token 限制
@@ -221,7 +223,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
         response = await process_chat(event, send_messages, tokens)
 
         # 记录模型回复
-        private_data["memory"]["messages"].append(
+        private_data.memory.messages.append(
             {
                 "role": "assistant",
                 "content": str(response),
@@ -234,7 +236,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
 
     async def manage_sessions(
         event: GroupMessageEvent | PrivateMessageEvent,
-        data: dict,
+        data: MemoryModel,
         session_clear_list: list,
     ):
         """
@@ -256,25 +258,25 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
                         break
 
                 # 检查会话超时
-                if (time.time() - data["timestamp"]) >= (
+                if (time.time() - data.timestamp) >= (
                     float(config_manager.config.session_control_time * 60)
                 ):
-                    data["sessions"].append(
+                    data.sessions.append(
                         {
-                            "messages": data["memory"]["messages"],
+                            "messages": data.memory.messages,
                             "time": time.time(),
                         }
                     )
                     while (
-                        len(data["sessions"])
+                        len(data.sessions)
                         > config_manager.config.session_control_history
                     ):
-                        data["sessions"].remove(data["sessions"][0])
-                    data["memory"]["messages"] = []
-                    data["timestamp"] = time.time()
+                        data.sessions.remove(data.sessions[0])
+                    data.memory.messages = []
+                    data.timestamp = time.time()
                     await write_memory_data(event, data)
                     if not (
-                        (time.time() - data["timestamp"])
+                        (time.time() - data.timestamp)
                         > float(config_manager.config.session_control_time * 60 * 2)
                     ):
                         chated = await matcher.send(
@@ -309,16 +311,14 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
                                         message_id=session["message_id"]
                                     )
                             session_clear_list.remove(session)
-                            data["memory"]["messages"] = data["sessions"][-1][
-                                "messages"
-                            ]
-                            data["sessions"].pop()
+                            data.memory.messages = data.sessions[-1]["messages"]
+                            data.sessions.pop()
                             await matcher.send("让我们继续聊天吧～")
                             await write_memory_data(event, data)
                             raise CancelException()
 
             finally:
-                data["timestamp"] = time.time()
+                data.timestamp = time.time()
 
     async def handle_reply(
         reply: Reply, bot: Bot, group_id: int | None, content: str
@@ -350,7 +350,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
             role, "[获取身份失败]"
         )
 
-    def enforce_memory_limit(data: dict, memory_length_limit: int):
+    def enforce_memory_limit(data: MemoryModel, memory_length_limit: int):
         """
         控制记忆长度，删除超出限制的旧消息，移除不支持的消息。
         """
@@ -358,7 +358,7 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
             config_manager.config.preset, fix=True
         )
         # Process multimodal messages when needed
-        for message in data["memory"]["messages"]:
+        for message in data.memory.messages:
             if (
                 isinstance(message["content"], dict)
                 and not is_multimodal
@@ -372,17 +372,17 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
 
         # Enforce memory length limit
         while (
-            len(data["memory"]["messages"]) > memory_length_limit
-            or (data["memory"]["messages"][0]["role"] != "user")
-        ) and len(data["memory"]["messages"]) > 0:
-            del data["memory"]["messages"][0]
+            len(data.memory.messages) > memory_length_limit
+            or (data.memory.messages[0]["role"] != "user")
+        ) and len(data.memory.messages) > 0:
+            del data.memory.messages[0]
 
-    async def enforce_token_limit(data: dict, train: dict) -> int:
+    async def enforce_token_limit(data: MemoryModel, train: dict[str, Any]) -> int:
         """
         控制 token 数量，删除超出限制的旧消息，返回处理后的Tokens。
         """
         train = copy.deepcopy(train)
-        memory_l = [train, *copy.deepcopy(data["memory"]["messages"].copy())]  # type: list[dict]
+        memory_l = [train, *copy.deepcopy(data.memory.messages.copy())]  # type: list[dict]
         full_string = ""
         for st in memory_l:
             if not st.get("content"):
@@ -402,8 +402,8 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
             return tokens
         while tokens > config_manager.config.session_max_tokens:
             try:
-                if len(data["memory"]["messages"]) > 0:
-                    del data["memory"]["messages"][0]
+                if len(data.memory.messages) > 0:
+                    del data.memory.messages[0]
                 else:
                     logger.warning(
                         f"提示词大小过大！为{hybrid_token_count(train['content'])}>{config_manager.config.session_max_tokens}！"
@@ -442,15 +442,15 @@ async def chat(event: MessageEvent, matcher: Matcher, bot: Bot):
             )
         return tokens
 
-    def prepare_send_messages(data: dict, train: dict) -> list:
+    def prepare_send_messages(data: MemoryModel, train: dict[str, Any]) -> list:
         """
         准备发送给聊天模型的消息列表，包括系统提示词数据和上下文。
         """
         train = copy.deepcopy(train)
         train["content"] += (
-            f"\n以下是一些补充内容，如果与上面任何一条有冲突请忽略。\n{data.get('prompt', '无')}"
+            f"\n以下是一些补充内容，如果与上面任何一条有冲突请忽略。\n{data.prompt}"
         )
-        send_messages = data["memory"]["messages"].copy()
+        send_messages = data.memory.messages.copy()
         send_messages.insert(0, train)
         return send_messages
