+++ src/nonebot_plugin_suggarchat/handle/sessions.py
@@ -8,7 +8,7 @@ from nonebot.params import CommandArg
 
 from ..check_rule import is_group_admin_if_is_in_group
 from ..config import config_manager
-from ..utils import get_memory_data, write_memory_data
+from ..utils.memory import MemoryModel, get_memory_data, write_memory_data
 
 
 async def sessions(
@@ -18,23 +18,23 @@ async def sessions(
     if not await is_group_admin_if_is_in_group(event, bot):
         await matcher.finish("你没有权限执行此命令。")
 
-    async def display_sessions(data: dict) -> None:
+    async def display_sessions(data: MemoryModel) -> None:
         """显示历史会话列表"""
-        if not data.get("sessions"):
+        if not data.sessions:
             await matcher.finish("没有历史会话")
         message_content = "历史会话\n"
-        for index, msg in enumerate(data["sessions"]):
+        for index, msg in enumerate(data.sessions):
             message_content += f"编号：{index}) ：{msg['messages'][0]['content'][9:]}... 时间：{datetime.fromtimestamp(msg['time']).strftime('%Y-%m-%d %I:%M:%S %p')}\n"
         await matcher.finish(message_content)
 
-    async def set_session(data: dict, arg_list: list[str], event: MessageEvent) -> None:
+    async def set_session(
+        data: MemoryModel, arg_list: list[str], event: MessageEvent
+    ) -> None:
         """将当前会话覆盖为指定编号的会话"""
         try:
             if len(arg_list) >= 2:
-                data["memory"]["messages"] = data["sessions"][int(arg_list[1])][
-                    "messages"
-                ]
-                data["timestamp"] = time.time()
+                data.memory.messages = data.sessions[int(arg_list[1])]["messages"]
+                data.timestamp = time.time()
                 await write_memory_data(event, data)
                 await matcher.send("完成记忆覆盖。")
             else:
@@ -45,12 +45,12 @@ async def sessions(
             await matcher.finish("覆盖记忆文件失败，这个对话可能损坏了。")
 
     async def delete_session(
-        data: dict, arg_list: list[str], event: MessageEvent
+        data: MemoryModel, arg_list: list[str], event: MessageEvent
     ) -> None:
         """删除指定编号的会话"""
         try:
             if len(arg_list) >= 2:
-                data["sessions"].remove(data["sessions"][int(arg_list[1])])
+                data.sessions.remove(data.sessions[int(arg_list[1])])
                 await write_memory_data(event, data)
             else:
                 await matcher.finish("请输入正确编号")
@@ -59,18 +59,18 @@ async def sessions(
         except Exception:
             await matcher.finish("删除指定编号会话失败。")
 
-    async def archive_session(data: dict, event: MessageEvent) -> None:
+    async def archive_session(data: MemoryModel, event: MessageEvent) -> None:
         """归档当前会话"""
         try:
-            if data["memory"]["messages"]:
-                data["sessions"].append(
+            if data.memory.messages:
+                data.sessions.append(
                     {
-                        "messages": data["memory"]["messages"],
+                        "messages": data.memory.messages,
                         "time": time.time(),
                     }
                 )
-                data["memory"]["messages"] = []
-                data["timestamp"] = time.time()
+                data.memory.messages = []
+                data.timestamp = time.time()
                 await write_memory_data(event, data)
                 await matcher.finish("当前会话已归档。")
             else:
@@ -80,11 +80,11 @@ async def sessions(
         except Exception:
             await matcher.finish("归档当前会话失败。")
 
-    async def clear_sessions(data: dict, event: MessageEvent) -> None:
+    async def clear_sessions(data: MemoryModel, event: MessageEvent) -> None:
         """清空所有会话"""
         try:
-            data["sessions"] = []
-            data["timestamp"] = time.time()
+            data.sessions = []
+            data.timestamp = time.time()
             await write_memory_data(event, data)
             await matcher.finish("会话已清空。")
         except NoneBotException as e:
